+++ CHANGELOG.md
@@ -2,6 +2,42 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [4.9.3](https://github.com/streetsidesoftware/vscode-spell-checker/compare/code-spell-checker-v4.9.2...code-spell-checker-v4.9.3) (2026-09-11)
+
+
+### Updates and Bug Fixes
+
+* issue with missing menu items due to VSCode bug ([#5504](https://github.com/streetsidesoftware/vscode-spell-checker/issues/5504)) ([a163f4f](https://github.com/streetsidesoftware/vscode-spell-checker/commit/a163f4f2515d3ff307b99882660aaca9fa62ede5))
+
+## [4.9.2](https://github.com/streetsidesoftware/vscode-spell-checker/compare/code-spell-checker-v4.9.1...code-spell-checker-v4.9.2) (2026-09-10)
+
+
+### Updates and Bug Fixes
+
+* Prepare for release ([#5491](https://github.com/streetsidesoftware/vscode-spell-checker/issues/5491)) ([ae119e7](https://github.com/streetsidesoftware/vscode-spell-checker/commit/ae119e71bb2c98250a067cdfd89f97e023f658c6))
+* Prepare for Release ([#5497](https://github.com/streetsidesoftware/vscode-spell-checker/issues/5497)) ([a16ef89](https://github.com/streetsidesoftware/vscode-spell-checker/commit/a16ef89e0186521f45a05b1dbe6e63dd27562268))
+
+## [4.9.1](https://github.com/streetsidesoftware/vscode-spell-checker/compare/code-spell-checker-v4.9.0...code-spell-checker-v4.9.1) (2026-09-09)
+
+
+### Updates and Bug Fixes
+
+* Add control for menu items and actions ([#5482](https://github.com/streetsidesoftware/vscode-spell-checker/issues/5482)) ([9a513a1](https://github.com/streetsidesoftware/vscode-spell-checker/commit/9a513a17c8b51a5dc2556a92930e99fe802aaba7))
+* Be able to control the items on the action menu ([#5488](https://github.com/streetsidesoftware/vscode-spell-checker/issues/5488)) ([45888a3](https://github.com/streetsidesoftware/vscode-spell-checker/commit/45888a382f9db290278697920d91587f92089db3))
+* Use allowWordsToBeAddTo to filter Actions ([#5490](https://github.com/streetsidesoftware/vscode-spell-checker/issues/5490)) ([6cd0af7](https://github.com/streetsidesoftware/vscode-spell-checker/commit/6cd0af7d1c9e36e82a4a2771588c486e8a11e504))
+
+## [4.9.0](https://github.com/streetsidesoftware/vscode-spell-checker/compare/code-spell-checker-v4.8.0...code-spell-checker-v4.9.0) (2026-09-08)
+
+
+### Features
+
+* Update CSpell from `10.2.2` to `10.3.0` and other dependencies ([4c37798](https://github.com/streetsidesoftware/vscode-spell-checker/commit/4c37798a14a3fd2df8da8c042d59832eae33f054))
+
+
+### Updates and Bug Fixes
+
+* **website:** Improve display of default values ([#5464](https://github.com/streetsidesoftware/vscode-spell-checker/issues/5464)) ([4564618](https://github.com/streetsidesoftware/vscode-spell-checker/commit/45646189c2c5fb51bae5ffc15f5d5ee849ef32dd))
+
 ## [4.8.0](https://github.com/streetsidesoftware/vscode-spell-checker/compare/code-spell-checker-v4.7.5...code-spell-checker-v4.8.0) (2026-09-04)